ChatBot_LLM_Llama/requirements.txt
fastapi
uvicorn
llama-cpp-python
autoscraper
selenium
pydantic
twilio